chore(library): rename react helper files
