+++ packages/helper-react/src/useFieldPlugin.ts
@@ -1,6 +1,6 @@
 import { FieldPluginResponse } from '@storyblok/field-plugin'
 import { useContext } from 'react'
-import { FieldPluginContext } from './context'
+import { FieldPluginContext } from './FieldPluginContext'
 
 export const useFieldPlugin = (): Extract<
   FieldPluginResponse,
